tests: reduce instability of test_exitcode_previous_to_join on PyPy
On PyPy 5.6.0 on a T450s the test test_exitcode_previous_to_join
failed 50 out of 100 times. The test is intentionally racy as it
confirms that the event machinery picks up and processes a SIGCHLD
signal during a `gevent.sleep(T)`. This patch prolongs the sleep
duration by factor 5 (from 0.1 s to 0.5 s). This makes the test
pass 100 out of 100 times on said test platform. This instability
has never been observed with the CPython interpreter.