gipc: do not serialize lock when pickling handles
A `_GIPCHandle` lock state is only meaningful in the current
process when the handle is actively being used.

When a `_GIPCHandle` instance is being pickled it is fair to assume
that the lock state will not be meaningful anymore in the moment
when the handle is going to be unpickled again. Most likely,
then the lock state will even be invalid. One definite example
for this is when the handle is unpickled in a process different
from the process it was pickled in -- the lock implementation
cannot synchronize across process boundaries.

This patch hooks into the pickling and unpicklung process of
`_GIPCHandle` objects. It removes the lock object from the
serialized state before returning it to consumers, and it creates
a fresh lock upon deserialization.

This makes sense conceptually and addresses a gipc bug on Windows
with recent gevent versions where the (unnecessary) pickling of
a gevent.lock.Semaphore object fails with:

  File "C:\Anaconda\lib\multiprocessing\reduction.py", line 60, in dump
    ForkingPickler(file, protocol).dump(obj)
  TypeError: can't pickle gevent._semaphore.Semaphore objects