ci: take control of PYENV_ROOT environment variable
Travis CI sets this to /opt/pyenv which holds an
old pyenv release. For the pyenv installation from
current head of master to succeed we set PYENV_ROOT
to a directory that does not exist by default and
that the build job can write to. Once written it
is cached by Travis CI though.